chore: relocate react tests
