+++ apps/package/jest/react/WavelengthAppLogo.test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { WavelengthAppLogo } from "../src";
+import { WavelengthAppLogo } from "../../src";
 import { expect } from "@jest/globals";
 import { render, screen, fireEvent } from "@testing-library/react";
