+++ apps/package/jest/react/WavelengthBox.test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { WavelengthBox } from "../src";
+import { WavelengthBox } from "../../src";
 import { expect } from "@jest/globals";
 import { render, screen, fireEvent } from "@testing-library/react";
