+++ apps/package/jest/react/WavelengthButton.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen, fireEvent } from "@testing-library/react";
-import { WavelengthButton } from "../src";
+import { WavelengthButton } from "../../src";
 import { jest } from "@jest/globals";
 import "@testing-library/jest-dom";
