+++ apps/package/jest/react/WavelengthFooter.test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { WavelengthFooter } from "../src";
+import { WavelengthFooter } from "../../src";
 import { fireEvent, render, screen } from "@testing-library/react";
 
 describe("Wavelength Footer", () => {
