+++ apps/package/jest/react/WavelengthTitleBar.test.tsx
@@ -1,13 +1,13 @@
 import React from "react";
-import { SampleComponent } from "../src";
+import { WavelengthTitleBar } from "../../src";
 import { expect } from "@jest/globals";
 import { render, screen, fireEvent } from "@testing-library/react";
 
-describe("SampleComponent should", () => {
+describe("WavelengthTitleBar should", () => {
   it("render", async () => {
     render(
       <div data-testid="default">
-        <SampleComponent testProp="Sample Component" children="Sample Component" />
+        <WavelengthTitleBar titleText="Test" subtitleText="Test" textColor="hotpink" textShadow />
       </div>,
     );
     const titleBar = screen.getByTestId("default");
